requirements.txt
langchain==0.3.2
langchain-openai==0.2.2
langchain-community==0.3.1
langgraph==0.2.21
faiss-cpu==1.8.0
tiktoken==0.7.0
python-dotenv==1.0.1
pydantic==2.9.2
fastapi==0.115.2
uvicorn==0.30.1
streamlit==1.38.0
pypdf==4.3.1
duckduckgo-search==6.3.5
langsmith
